# chore: Optimize Filter() for small lists